requirements.txt
Django==1.6
Markdown==2.3.1
South==0.8.3
distribute==0.6.34
django-filter==0.7
djangorestframework==2.3.9
postgres==2.1.1
psycopg2==2.5.1
wsgiref==0.1.2
